Add IAM policy for AWS Product API Search
Officially, the Product API does not support IAM policies:

> IAM roles are not currently supported. You must use the root account credentials.

There are some reports that individuals have successfully used
an IAM policy by providing full access to everything in AWS
and then restricting the resources which it has access to.

The recommended approach is to use the root credentials for the
product API. This would work; however, it has the quality that
with the credentials an AWS account could be stolen or fully
comprimised.